refactor: 댓글 컴포넌트 분리(댓글 수정시 리랜더링 방지)
